feat(variables) :  - allow to use variables in variables
